Refactor stratis-utils for better modularity and more convenient dispatch
Implement dispatch in the same way as in stratisd-tools, but return an
ExecutableError from the main method as before. this allows invoking the
executables as subcommands of straits-utils as well as via links.

Move the generators module into the utils module.

Establish a file for the predict_usage functionality and copy the code
from stratis-utils to that file. Introduce a PredictError class to wrap
up string errors as they come along.

Implement two run methods for StratisClevisSetupGenerator and
StratisSetupGenerator to retain behavior on the executable not having
been compiled with systemd compatibility. In that case, it will
recognize the commands, but return an error indicating that systemd
support is not implemented.
