+++ src/bin/utils/generators/stratis_setup_generator.rs
