+++ src/bin/utils/mod.rs
@@ -1,0 +1,10 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+mod cmds;
+#[cfg(feature = "systemd_compat")]
+mod generators;
+mod predict_usage;
+
+pub use cmds::{cmds, ExecutableError};
